+++ include/rocksdb/iterator.h
@@ -19,6 +19,7 @@
 #ifndef STORAGE_ROCKSDB_INCLUDE_ITERATOR_H_
 #define STORAGE_ROCKSDB_INCLUDE_ITERATOR_H_
 
+#include <string>
 #include "rocksdb/slice.h"
 #include "rocksdb/status.h"
 
@@ -95,14 +96,16 @@ class Iterator : public Cleanable {
   // satisfied without doing some IO, then this returns Status::Incomplete().
   virtual Status status() const = 0;
 
-  // If true, this means that the Slice returned by key() is valid as long
-  // as the iterator is not deleted and ReleasePinnedData() is not called.
-  //
-  // IsKeyPinned() is guaranteed to always return true if
-  //  - Iterator created with ReadOptions::pin_data = true
-  //  - DB tables were created with BlockBasedTableOptions::use_delta_encoding
-  //    set to false.
-  virtual bool IsKeyPinned() const { return false; }
+  // Property "rocksdb.iterator.is.key.pinned":
+  //   If returning "1", this means that the Slice returned by key() is valid
+  //   as long as the iterator is not deleted and ReleasePinnedData() is not
+  //   called.
+  //   It is guaranteed to always return "1" if
+  //      - Iterator created with ReadOptions::pin_data = true
+  //      - DB tables were created with
+  //      BlockBasedTableOptions::use_delta_encoding
+  //        set to false.
+  virtual Status GetProperty(std::string prop_name, std::string* prop);
 
  private:
   // No copying allowed